@@ -1,31 +1,15 @@
-/*import React from 'react';
-
-const ContactList = ({ contacts, handleDeleteContact }) => {
-  return (
-    <ul>
-      {contacts.map((contact) => (
-        <li key={contact.id}>
-          {contact.name} - {contact.number}
-          <button onClick={() => handleDeleteContact(contact.id)}>Delete</button>
-        </li>
-      ))}
-    </ul>
-  );
-};
-
-export default ContactList;*/
-
-
 import React from 'react';
 import PropTypes from 'prop-types';
 
 const ContactList = ({ contacts, handleDeleteContact }) => {
   return (
     <ul>
-      {contacts.map((contact) => (
+      {contacts.map(contact => (
         <li key={contact.id}>
           {contact.name} - {contact.number}
-          <button onClick={() => handleDeleteContact(contact.id)}>Delete</button>
+          <button onClick={() => handleDeleteContact(contact.id)}>
+            Delete
+          </button>
         </li>
       ))}
     </ul>
@@ -44,5 +28,3 @@ ContactList.propTypes = {
 };
 
 export default ContactList;
-
-
